refactor: update instruction index type to u64
Update the instruction index type from u8 to u64 in the `CreateSlashEvent` struct and related functions. This change allows for a larger range of indices to be used for slash events.

Refactor the code to use `index.to_be_bytes()` when generating the seed for the slash event address in the `get_slash_event_address` function.

This commit addresses the need to support a larger number of slash events and ensures consistency in the codebase.

+++ staking/programs/integrity-pool/src/state/pool.rs
@@ -405,7 +405,7 @@ pub struct PoolConfig {
     pub reward_program_authority: Pubkey,
     pub pyth_token_mint:          Pubkey,
     pub y:                        frac64,
-    pub num_slash_events:         u8,
+    pub num_slash_events:         u64,
 }
 
 impl PoolConfig {
